# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sonart Insight API",
    "description": "\nERP ve reklam raporlarini yonetim dashboard'una ceviren analiz servisi.\n\n**Mimarinin iki tasiyici fikri:**\n\n1. **Dataset Pack** — boru hattinin tamami (okuma, kodlama onarimi, veri\n   kalitesi, zaman serisi, risk tespiti, AI) rapor tipinden bagimsizdir. Bir\n   rapora ozgu her sey tek bir yapilandirma nesnesinde toplanir. Yeni rapor\n   eklemek = yeni pack yazmak.\n\n2. **Sayiyi Polars hesaplar, AI yorumlar.** Modele ham veri hic verilmez;\n   yalnizca hesaplanmis tablolar gider. Uretilen her aksiyon sema geregi kanit\n   tasimak zorundadir ve her kanit hesaplanmis degerle karsilastirilarak\n   dogrulanir (`grounding_ratio`).\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/datasets/{dataset_id}": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Veri seti özeti",
        "operationId": "get_dataset_api_v1_datasets__dataset_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetSummary"
                }
              }
            }
          },
          "400": {
            "description": "İstek geçersiz (bilinmeyen rapor tipi, hatalı parametre)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Dosya işlenemedi ya da doğrulama hatası",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Beklenmeyen hata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Dosya boyutu izin verilen sınırı aşıyor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "datasets"
        ],
        "summary": "Veri setini sil",
        "operationId": "delete_dataset_api_v1_datasets__dataset_id__delete",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "İstek geçersiz (bilinmeyen rapor tipi, hatalı parametre)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Dosya işlenemedi ya da doğrulama hatası",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Beklenmeyen hata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Dosya boyutu izin verilen sınırı aşıyor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/datasets/{dataset_id}/analysis/status": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Analiz hazır mı, kaç çağrı gerekecek",
        "description": "Frontend'in 'analiz üret' butonunu maliyet uyarısıyla göstermesi için. Önbellekte varsa çağrı yapılmayacağını bildirir.",
        "operationId": "analysis_status_api_v1_datasets__dataset_id__analysis_status_get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "İstek geçersiz (bilinmeyen rapor tipi, hatalı parametre)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Dosya işlenemedi ya da doğrulama hatası",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Beklenmeyen hata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "AI servisi istek limitini aştı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "AI servisinden geçerli yanıt alınamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "AI yapılandırılmamış (ANTHROPIC_API_KEY yok)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/analysis": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Dönemsel analiz + yönetici özeti üret",
        "description": "Her dönem için paralel bir AI çağrısı, ardından tek bir sentez çağrısı. Sonuç önbelleğe yazılır; aynı veri seti için tekrar çağrıldığında AI'a gidilmez. `refresh=true` ile yeniden üretilebilir.",
        "operationId": "create_analysis_api_v1_datasets__dataset_id__analysis_post",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Önbelleği yok say ve yeniden üret",
              "default": false,
              "title": "Refresh"
            },
            "description": "Önbelleği yok say ve yeniden üret"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "İstek geçersiz (bilinmeyen rapor tipi, hatalı parametre)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Dosya işlenemedi ya da doğrulama hatası",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Beklenmeyen hata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "AI servisi istek limitini aştı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "AI servisinden geçerli yanıt alınamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "AI yapılandırılmamış (ANTHROPIC_API_KEY yok)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/ask": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Veri üzerinde serbest soru-cevap",
        "description": "Model SQL yazmaz: hesaplanmış tabloları yorumlar. Böylece 'AI sayı hesaplamaz' kuralı soru-cevapta da korunur ve her yanıt kanıt taşır.",
        "operationId": "ask_api_v1_datasets__dataset_id__ask_post",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            }
          },
          "400": {
            "description": "İstek geçersiz (bilinmeyen rapor tipi, hatalı parametre)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Kayıt bulunamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Dosya işlenemedi ya da doğrulama hatası",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Beklenmeyen hata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "AI servisi istek limitini aştı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "AI servisinden geçerli yanıt alınamadı",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "AI yapılandırılmamış (ANTHROPIC_API_KEY yok)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Action": {
        "properties": {
          "priority": {
            "type": "string",
            "enum": [
              "kritik",
              "yuksek",
              "orta",
              "dusuk"
            ],
            "title": "Priority"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Tek cumlelik, emir kipinde eylem basligi."
          },
          "rationale": {
            "type": "string",
            "title": "Rationale",
            "description": "Neden simdi ve neden bu urun/kayit icin."
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "minItems": 1,
            "title": "Evidence",
            "description": "Bu aksiyonu dogrulayan, tablolardan alinmis sayilar."
          },
          "owner": {
            "type": "string",
            "title": "Owner",
            "description": "Sorumlu departman."
          },
          "horizon": {
            "type": "string",
            "enum": [
              "bu_hafta",
              "bu_ay",
              "bu_ceyrek"
            ],
            "title": "Horizon"
          },
          "expected_impact_tl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Impact Tl",
            "description": "Parasal etki tahmini; hesaplanamiyorsa bos birak."
          }
        },
        "type": "object",
        "required": [
          "priority",
          "title",
          "rationale",
          "evidence",
          "owner",
          "horizon"
        ],
        "title": "Action",
        "description": "Somut, sahiplenilebilir bir aksiyon onerisi."
      },
      "AnalysisResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "pack_key": {
            "type": "string",
            "title": "Pack Key"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "description": "True ise sonuç önbellekten geldi; yeni AI çağrısı yapılmadı."
          },
          "cached_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cached At"
          },
          "periods": {
            "items": {
              "$ref": "#/components/schemas/PeriodAnalysis"
            },
            "type": "array",
            "title": "Periods"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutiveSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "grounding": {
            "$ref": "#/components/schemas/GroundingOut"
          },
          "telemetry": {
            "$ref": "#/components/schemas/TelemetryOut"
          },
          "failed_periods": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Failed Periods"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "pack_key",
          "model",
          "cached",
          "periods",
          "summary",
          "grounding",
          "telemetry"
        ],
        "title": "AnalysisResponse"
      },
      "AnalysisStatusResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "ai_configured": {
            "type": "boolean",
            "title": "Ai Configured"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "effort": {
            "type": "string",
            "title": "Effort"
          },
          "estimated_calls": {
            "type": "integer",
            "title": "Estimated Calls"
          },
          "detail": {
            "type": "object",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "ai_configured",
          "cached",
          "model",
          "effort",
          "estimated_calls"
        ],
        "title": "AnalysisStatusResponse"
      },
      "AskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Question",
            "examples": [
              "Marjı en hızlı daralan ürün hangisi?"
            ]
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AskRequest"
      },
      "AskResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "answer": {
            "$ref": "#/components/schemas/QAAnswer"
          },
          "grounding": {
            "$ref": "#/components/schemas/GroundingOut"
          },
          "telemetry": {
            "$ref": "#/components/schemas/TelemetryOut"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "dataset_id",
          "question",
          "answer",
          "grounding",
          "telemetry"
        ],
        "title": "AskResponse"
      },
...
      "DatasetSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "pack_key": {
            "type": "string",
            "title": "Pack Key"
          },
          "pack_title": {
            "type": "string",
            "title": "Pack Title"
          },
          "raw_row_count": {
            "type": "integer",
            "title": "Raw Row Count"
          },
          "clean_row_count": {
            "type": "integer",
            "title": "Clean Row Count"
          },
          "quarantined_row_count": {
            "type": "integer",
            "title": "Quarantined Row Count"
          },
          "health_score": {
            "type": "integer",
            "title": "Health Score"
          },
          "encoding_detected": {
            "type": "string",
            "title": "Encoding Detected"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "pack_key",
          "pack_title",
          "raw_row_count",
          "clean_row_count",
          "quarantined_row_count",
          "health_score",
          "encoding_detected",
          "created_at"
        ],
        "title": "DatasetSummary"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "schema_mismatch"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Tum hatalarin ortak govdesi.\n\nIstemci `code` alanina gore dallanir; mesaj metni degisse bile kirilmaz."
      },
      "Evidence": {
        "properties": {
          "metric": {
            "type": "string",
            "title": "Metric",
            "description": "Metrigin teknik adi, or. 'son_kapama_ay'."
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Verilen tablolardaki degerin aynisi."
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "description": "Birim: TL, adet, %, ay, oran, sayi."
          },
          "entity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity",
            "description": "Degerin ait oldugu kapsam. Tek bir kayit ise kodu ('U005', 'K011'); bir kategori/depo/platform kirilimi ise o boyutun adi ('Cantalik', 'Yan Depo'); yalnizca portfoyun tamamiysa bos birak. Kirilim degerini bos birakma -- o zaman portfoy geneliyle karisir."
          },
          "period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period",
            "description": "Donem, or. '2026-03'."
          }
        },
        "type": "object",
        "required": [
          "metric",
          "value",
          "unit"
        ],
        "title": "Evidence",
        "description": "Bir iddianin dayandigi somut sayi."
      },
      "ExecutiveSummary": {
        "properties": {
          "headline": {
            "type": "string",
            "title": "Headline",
            "description": "Yonetimin ilk okuyacagi tek cumle."
          },
          "situation": {
            "type": "string",
            "title": "Situation",
            "description": "Genel durum: 2-4 cumle, sayiya dayali."
          },
          "period_narrative": {
            "type": "string",
            "title": "Period Narrative",
            "description": "Alti donemin hikayesi: nereden nereye gelindi, kirilma noktalari hangi donemlerde yasandi."
          },
          "top_risks": {
            "items": {
              "$ref": "#/components/schemas/RiskHighlight"
            },
            "type": "array",
            "title": "Top Risks",
            "description": "En kritik riskler, oncelik sirasiyla."
          },
          "opportunities": {
            "items": {
              "$ref": "#/components/schemas/RiskHighlight"
            },
            "type": "array",
            "title": "Opportunities",
            "description": "Degerlendirilmesi gereken firsatlar."
          },
          "strategic_actions": {
            "items": {
              "$ref": "#/components/schemas/Action"
            },
            "type": "array",
            "title": "Strategic Actions",
            "description": "Yonetim seviyesinde stratejik aksiyonlar."
          }
        },
        "type": "object",
        "required": [
          "headline",
          "situation",
          "period_narrative",
          "top_risks",
          "strategic_actions"
        ],
        "title": "ExecutiveSummary",
        "description": "Yonetime hitap eden ust duzey sentez."
      },
      "GroundingOut": {
        "properties": {
          "total_evidence": {
            "type": "integer",
            "title": "Total Evidence"
          },
          "verified_evidence": {
            "type": "integer",
            "title": "Verified Evidence"
          },
          "grounding_ratio": {
            "type": "number",
            "title": "Grounding Ratio"
          },
          "issues": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "total_evidence",
          "verified_evidence",
          "grounding_ratio"
        ],
        "title": "GroundingOut",
        "description": "AI ciktisinin ne kadarinin hesaplanmis veriye dayandigi."
      },
...
      "PeriodAnalysis": {
        "properties": {
          "period": {
            "type": "string",
            "title": "Period"
          },
          "headline": {
            "type": "string",
            "title": "Headline",
            "description": "Bu donemi tek cumlede ozetleyen baslik."
          },
          "delta_vs_prev": {
            "type": "string",
            "title": "Delta Vs Prev",
            "description": "Bir onceki doneme gore somut olarak ne degisti. Ilk donemde baslangic durumunu tarif et. Genel ifade degil, sayiya dayali fark."
          },
          "dominant_dynamics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dominant Dynamics",
            "description": "Bu donemi belirleyen dinamikler (verilen listeden secilir)."
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/Action"
            },
            "type": "array",
            "title": "Actions",
            "description": "Bu doneme ozgu aksiyonlar."
          },
          "watch_items": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Watch Items",
            "description": "Henuz aksiyon gerektirmeyen ama izlenmesi gerekenler."
          }
        },
        "type": "object",
        "required": [
          "period",
          "headline",
          "delta_vs_prev",
          "dominant_dynamics",
          "actions"
        ],
        "title": "PeriodAnalysis",
        "description": "Tek bir donemin analizi.\n\n`delta_vs_prev` bilerek zorunlu: modelin her donem icin ayni genel metni\nuretmesini engelleyip bir onceki doneme gore *ne degistigini* soylemeye\nmecbur birakir. Donemsel farklilasmanin sema tarafindaki garantisi budur."
      },
...
      "QAAnswer": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer",
            "description": "Soruya dogrudan, Turkce yanit."
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence",
            "description": "Yaniti destekleyen, tablolardan alinmis sayilar."
          },
          "confidence": {
            "type": "string",
            "enum": [
              "yuksek",
              "orta",
              "dusuk"
            ],
            "title": "Confidence"
          },
          "caveats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Caveats",
            "description": "Yanitin sinirlari; veri yetersizse veya soru kapsam disiysa burada belirt."
          }
        },
        "type": "object",
        "required": [
          "answer",
          "confidence"
        ],
        "title": "QAAnswer",
        "description": "Serbest soru-cevap yaniti."
      },
...
      "RiskHighlight": {
        "properties": {
          "entity": {
            "type": "string",
            "title": "Entity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "why_it_matters": {
            "type": "string",
            "title": "Why It Matters"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "minItems": 1,
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "entity",
          "title",
          "why_it_matters",
          "evidence"
        ],
        "title": "RiskHighlight"
      },
...
      "TelemetryOut": {
        "properties": {
          "call_count": {
            "type": "integer",
            "title": "Call Count",
            "default": 0
          },
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens",
            "default": 0
          },
          "cache_write_tokens": {
            "type": "integer",
            "title": "Cache Write Tokens",
            "default": 0
          },
          "cache_read_tokens": {
            "type": "integer",
            "title": "Cache Read Tokens",
            "default": 0
          },
          "cache_hit_calls": {
            "type": "integer",
            "title": "Cache Hit Calls",
            "default": 0
          },
          "total_cost_usd": {
            "type": "number",
            "title": "Total Cost Usd",
            "default": 0.0
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms",
            "default": 0
          }
        },
        "type": "object",
        "title": "TelemetryOut"
      },
...
    }
  }
}